requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
gitpython==3.1.46
tree-sitter==0.25.2
chromadb==1.5.5
sentence-transformers==5.3.0
networkx==3.4.2
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
httpx==0.28.1
aiofiles==25.1.0
jinja2==3.1.6
python-multipart==0.0.22
